+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added more verbose logging feature.
+  [#12](https://github.com/OSC/mod_ood_proxy/issues/12)
 
 ## [0.3.1] - 2017-07-25
 ### Fixed